requirements.txt
# I recommend installing within a conda environment by using miniconda.
# See https://docs.conda.io/en/main/miniconda.html

ansible
ansible-pylibssh
ansible-runner>=2.2.1
dnacentersdk
flake8
jupyterlab
ipywidgets
nest-asyncio
orionsdk
pan-python
pandas
pan-os-python
python-dotenv[cli]
simplejson
pip
infoblox-client
meraki
pynetbox
python-dotenv
python-nmap
xmltodict
streamlit
tabulate
urllib3>=1.26.13,<=1.26.155
watchdog
